Bring across cvs from -current (which is at about version 1.6.3). This has remote repository access via "rsh". (eg: $CVSROOT = freefall.freebsd.org:/home/ncvs)
This is not quite the latest version, but it's a "known quantity".

There were a few covert tagging operations done as well to bring some
new files onto the -stable branch, these do not show here.

+++ gnu/usr.bin/cvs/lib/argmatch.c
@@ -21,6 +21,8 @@
 #include "config.h"
 #endif
 
+#include <sys/types.h>
+
 #include <stdio.h>
 #ifdef STDC_HEADERS
 #include <string.h>
@@ -39,12 +41,12 @@ argmatch (arg, optlist)
      char **optlist;
 {
   int i;			/* Temporary index in OPTLIST. */
-  int arglen;			/* Length of ARG. */
+  size_t arglen;		/* Length of ARG. */
   int matchind = -1;		/* Index of first nonexact match. */
   int ambiguous = 0;		/* If nonzero, multiple nonexact match(es). */
-
+  
   arglen = strlen (arg);
-
+  
   /* Test all elements for either exact match or abbreviated matches.  */
   for (i = 0; optlist[i]; i++)
     {
